add generated Scala2/Scala3 code from higher order functions before implementation
