+++ scala2/src/main/scala/examples/higher_order_fun.scala
@@ -1,0 +1,31 @@
+package examples
+
+object higher_order_fun
+{
+
+def id[A](x1: A): A = x1
+
+sealed trait Maybe[+A]
+object Maybe {
+  final case class Just[+A](x0: A) extends Maybe[A]
+  case object None extends Maybe[Nothing]
+}
+
+sealed trait List[+X]
+object List {
+  case object Nil extends List[Nothing]
+  final case class Cons[+X](x0: X, x1: List[X]) extends List[X]
+}
+
+def emptyNatList(): List[Long] = List.Nil
+
+/* TODO In printScala2 got SeUnhandled "QName {qnameModule = MName {mnameToList = [Name {nameId = NameId 0 (ModuleNameHash 2665135241066394892), nameConcrete = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"examples\" :| []}, nameCanonical = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"examples\" :| []}, nameBindingSite = Range (Just (RangeFile {rangeFilePath = AbsolutePath {textPath = \"/home/piter/agda2scala/examples/higher_order_fun.agda\"}, rangeFileName = Just (TopLevelModuleName {moduleNameRange = NoRange, moduleNameId = ModuleNameHash 2665135241066394892, moduleNameParts = \"examples\" :| [\"higher_order_fun\"], moduleNameInferred = False})})) (fromList [Interval {getIntervalFile = (), iStart' = Pn {srcFile = (), posPos = 8, posLine = 1, posCol = 8}, iEnd' = Pn {srcFile = (), posPos = 17, posLine = 1, posCol = 17}}]), nameFixity = Fixity' {theFixity = Fixity {fixityRange = NoRange, fixityLevel = Unrelated, fixityAssoc = NonAssoc}, theNotation = [], theNameRange = NoRange}, nameIsRecordName = False},Name {nameId = NameId 2 (ModuleNameHash 2665135241066394892), nameConcrete = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"higher\" :| [Hole,Id \"order\",Hole,Id \"fun\"]}, nameCanonical = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"higher\" :| [Hole,Id \"order\",Hole,Id \"fun\"]}, nameBindingSite = Range (Just (RangeFile {rangeFilePath = AbsolutePath {textPath = \"/home/piter/agda2scala/examples/higher_order_fun.agda\"}, rangeFileName = Just (TopLevelModuleName {moduleNameRange = NoRange, moduleNameId = ModuleNameHash 2665135241066394892, moduleNameParts = \"examples\" :| [\"higher_order_fun\"], moduleNameInferred = False})})) (fromList [Interval {getIntervalFile = (), iStart' = Pn {srcFile = (), posPos = 17, posLine = 1, posCol = 17}, iEnd' = Pn {srcFile = (), posPos = 33, posLine = 1, posCol = 33}}]), nameFixity = Fixity' {theFixity = Fixity {fixityRange = NoRange, fixityLevel = Unrelated, fixityAssoc = NonAssoc}, theNotation = [], theNameRange = NoRange}, nameIsRecordName = False}]}, qnameName = Name {nameId = NameId 38 (ModuleNameHash 2665135241066394892), nameConcrete = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"map\" :| []}, nameCanonical = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"map\" :| []}, nameBindingSite = Range (Just (RangeFile {rangeFilePath = AbsolutePath {textPath = \"/home/piter/agda2scala/examples/higher_order_fun.agda\"}, rangeFileName = Just (TopLevelModuleName {moduleNameRange = NoRange, moduleNameId = ModuleNameHash 2665135241066394892, moduleNameParts = \"examples\" :| [\"higher_order_fun\"], moduleNameInferred = False})})) (fromList [Interval {getIntervalFile = (), iStart' = Pn {srcFile = (), posPos = 549, posLine = 28, posCol = 1}, iEnd' = Pn {srcFile = (), posPos = 552, posLine = 28, posCol = 4}}]), nameFixity = Fixity' {theFixity = Fixity {fixityRange = NoRange, fixityLevel = Unrelated, fixityAssoc = NonAssoc}, theNotation = [], theNameRange = NoRange}, nameIsRecordName = False}}" "UnsupportedTerm (Pi (Dom {domInfo = ArgInfo {argInfoHiding = NotHidden, argInfoModality = Modality {modRelevance = Relevant ORelInferred, modQuantity = Quantity\969 Q\969Inferred, modCohesion = Continuous, modPolarity = PolarityModality {modPolarityAnn = MixedPolarity, modPolarityOrigin = MixedPolarity, modPolarityLock = StrictlyPositive}}, argInfoOrigin = UserWritten, argInfoFreeVariables = UnknownFVs, argInfoAnnotation = Annotation {annLock = IsNotLock}}, domName = Nothing, domIsFinite = False, domTactic = Nothing, unDom = El {_getSort = Univ UType (Max 0 []), unEl = Var 1 []}}) (NoAbs \"_\" El {_getSort = Univ UType (Max 0 []), unEl = Var 0 []}))" */
+
+/* TODO In printScala2 got SeUnhandled "QName {qnameModule = MName {mnameToList = [Name {nameId = NameId 0 (ModuleNameHash 2665135241066394892), nameConcrete = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"examples\" :| []}, nameCanonical = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"examples\" :| []}, nameBindingSite = Range (Just (RangeFile {rangeFilePath = AbsolutePath {textPath = \"/home/piter/agda2scala/examples/higher_order_fun.agda\"}, rangeFileName = Just (TopLevelModuleName {moduleNameRange = NoRange, moduleNameId = ModuleNameHash 2665135241066394892, moduleNameParts = \"examples\" :| [\"higher_order_fun\"], moduleNameInferred = False})})) (fromList [Interval {getIntervalFile = (), iStart' = Pn {srcFile = (), posPos = 8, posLine = 1, posCol = 8}, iEnd' = Pn {srcFile = (), posPos = 17, posLine = 1, posCol = 17}}]), nameFixity = Fixity' {theFixity = Fixity {fixityRange = NoRange, fixityLevel = Unrelated, fixityAssoc = NonAssoc}, theNotation = [], theNameRange = NoRange}, nameIsRecordName = False},Name {nameId = NameId 2 (ModuleNameHash 2665135241066394892), nameConcrete = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"higher\" :| [Hole,Id \"order\",Hole,Id \"fun\"]}, nameCanonical = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"higher\" :| [Hole,Id \"order\",Hole,Id \"fun\"]}, nameBindingSite = Range (Just (RangeFile {rangeFilePath = AbsolutePath {textPath = \"/home/piter/agda2scala/examples/higher_order_fun.agda\"}, rangeFileName = Just (TopLevelModuleName {moduleNameRange = NoRange, moduleNameId = ModuleNameHash 2665135241066394892, moduleNameParts = \"examples\" :| [\"higher_order_fun\"], moduleNameInferred = False})})) (fromList [Interval {getIntervalFile = (), iStart' = Pn {srcFile = (), posPos = 17, posLine = 1, posCol = 17}, iEnd' = Pn {srcFile = (), posPos = 33, posLine = 1, posCol = 33}}]), nameFixity = Fixity' {theFixity = Fixity {fixityRange = NoRange, fixityLevel = Unrelated, fixityAssoc = NonAssoc}, theNotation = [], theNameRange = NoRange}, nameIsRecordName = False}]}, qnameName = Name {nameId = NameId 52 (ModuleNameHash 2665135241066394892), nameConcrete = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"foldr\" :| []}, nameCanonical = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"foldr\" :| []}, nameBindingSite = Range (Just (RangeFile {rangeFilePath = AbsolutePath {textPath = \"/home/piter/agda2scala/examples/higher_order_fun.agda\"}, rangeFileName = Just (TopLevelModuleName {moduleNameRange = NoRange, moduleNameId = ModuleNameHash 2665135241066394892, moduleNameParts = \"examples\" :| [\"higher_order_fun\"], moduleNameInferred = False})})) (fromList [Interval {getIntervalFile = (), iStart' = Pn {srcFile = (), posPos = 687, posLine = 33, posCol = 1}, iEnd' = Pn {srcFile = (), posPos = 692, posLine = 33, posCol = 6}}]), nameFixity = Fixity' {theFixity = Fixity {fixityRange = NoRange, fixityLevel = Unrelated, fixityAssoc = NonAssoc}, theNotation = [], theNameRange = NoRange}, nameIsRecordName = False}}" "UnsupportedTerm (Pi (Dom {domInfo = ArgInfo {argInfoHiding = NotHidden, argInfoModality = Modality {modRelevance = Relevant ORelInferred, modQuantity = Quantity\969 Q\969Inferred, modCohesion = Continuous, modPolarity = PolarityModality {modPolarityAnn = MixedPolarity, modPolarityOrigin = MixedPolarity, modPolarityLock = StrictlyPositive}}, argInfoOrigin = UserWritten, argInfoFreeVariables = UnknownFVs, argInfoAnnotation = Annotation {annLock = IsNotLock}}, domName = Nothing, domIsFinite = False, domTactic = Nothing, unDom = El {_getSort = Univ UType (Max 0 []), unEl = Var 1 []}}) (NoAbs \"_\" El {_getSort = Univ UType (Max 0 []), unEl = Pi (Dom {domInfo = ArgInfo {argInfoHiding = NotHidden, argInfoModality = Modality {modRelevance = Relevant ORelInferred, modQuantity = Quantity\969 Q\969Inferred, modCohesion = Continuous, modPolarity = PolarityModality {modPolarityAnn = MixedPolarity, modPolarityOrigin = MixedPolarity, modPolarityLock = StrictlyPositive}}, argInfoOrigin = UserWritten, argInfoFreeVariables = UnknownFVs, argInfoAnnotation = Annotation {annLock = IsNotLock}}, domName = Nothing, domIsFinite = False, domTactic = Nothing, unDom = El {_getSort = Univ UType (Max 0 []), unEl = Var 0 []}}) (NoAbs \"_\" El {_getSort = Univ UType (Max 0 []), unEl = Var 0 []})}))" */
+
+/* TODO In printScala2 got SeUnhandled "QName {qnameModule = MName {mnameToList = [Name {nameId = NameId 0 (ModuleNameHash 2665135241066394892), nameConcrete = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"examples\" :| []}, nameCanonical = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"examples\" :| []}, nameBindingSite = Range (Just (RangeFile {rangeFilePath = AbsolutePath {textPath = \"/home/piter/agda2scala/examples/higher_order_fun.agda\"}, rangeFileName = Just (TopLevelModuleName {moduleNameRange = NoRange, moduleNameId = ModuleNameHash 2665135241066394892, moduleNameParts = \"examples\" :| [\"higher_order_fun\"], moduleNameInferred = False})})) (fromList [Interval {getIntervalFile = (), iStart' = Pn {srcFile = (), posPos = 8, posLine = 1, posCol = 8}, iEnd' = Pn {srcFile = (), posPos = 17, posLine = 1, posCol = 17}}]), nameFixity = Fixity' {theFixity = Fixity {fixityRange = NoRange, fixityLevel = Unrelated, fixityAssoc = NonAssoc}, theNotation = [], theNameRange = NoRange}, nameIsRecordName = False},Name {nameId = NameId 2 (ModuleNameHash 2665135241066394892), nameConcrete = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"higher\" :| [Hole,Id \"order\",Hole,Id \"fun\"]}, nameCanonical = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"higher\" :| [Hole,Id \"order\",Hole,Id \"fun\"]}, nameBindingSite = Range (Just (RangeFile {rangeFilePath = AbsolutePath {textPath = \"/home/piter/agda2scala/examples/higher_order_fun.agda\"}, rangeFileName = Just (TopLevelModuleName {moduleNameRange = NoRange, moduleNameId = ModuleNameHash 2665135241066394892, moduleNameParts = \"examples\" :| [\"higher_order_fun\"], moduleNameInferred = False})})) (fromList [Interval {getIntervalFile = (), iStart' = Pn {srcFile = (), posPos = 17, posLine = 1, posCol = 17}, iEnd' = Pn {srcFile = (), posPos = 33, posLine = 1, posCol = 33}}]), nameFixity = Fixity' {theFixity = Fixity {fixityRange = NoRange, fixityLevel = Unrelated, fixityAssoc = NonAssoc}, theNotation = [], theNameRange = NoRange}, nameIsRecordName = False}]}, qnameName = Name {nameId = NameId 68 (ModuleNameHash 2665135241066394892), nameConcrete = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"filter\" :| []}, nameCanonical = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"filter\" :| []}, nameBindingSite = Range (Just (RangeFile {rangeFilePath = AbsolutePath {textPath = \"/home/piter/agda2scala/examples/higher_order_fun.agda\"}, rangeFileName = Just (TopLevelModuleName {moduleNameRange = NoRange, moduleNameId = ModuleNameHash 2665135241066394892, moduleNameParts = \"examples\" :| [\"higher_order_fun\"], moduleNameInferred = False})})) (fromList [Interval {getIntervalFile = (), iStart' = Pn {srcFile = (), posPos = 856, posLine = 38, posCol = 1}, iEnd' = Pn {srcFile = (), posPos = 862, posLine = 38, posCol = 7}}]), nameFixity = Fixity' {theFixity = Fixity {fixityRange = NoRange, fixityLevel = Unrelated, fixityAssoc = NonAssoc}, theNotation = [], theNameRange = NoRange}, nameIsRecordName = False}}" "UnsupportedTerm (Pi (Dom {domInfo = ArgInfo {argInfoHiding = NotHidden, argInfoModality = Modality {modRelevance = Relevant ORelInferred, modQuantity = Quantity\969 Q\969Inferred, modCohesion = Continuous, modPolarity = PolarityModality {modPolarityAnn = MixedPolarity, modPolarityOrigin = MixedPolarity, modPolarityLock = StrictlyPositive}}, argInfoOrigin = UserWritten, argInfoFreeVariables = UnknownFVs, argInfoAnnotation = Annotation {annLock = IsNotLock}}, domName = Nothing, domIsFinite = False, domTactic = Nothing, unDom = El {_getSort = Univ UType (Max 0 []), unEl = Var 0 []}}) (NoAbs \"_\" El {_getSort = Univ UType (Max 0 []), unEl = Def (QName {qnameModule = MName {mnameToList = [Name {nameId = NameId 0 (ModuleNameHash 4305008439024043551), nameConcrete = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"Agda\" :| []}, nameCanonical = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"Agda\" :| []}, nameBindingSite = Range (Just (RangeFile {rangeFilePath = AbsolutePath {textPath = \"/home/piter/.cabal/store/ghc-9.12.2-9b33/Agda-2.8.0-7cf9474c14d743910af4cb5c11fc608c8aef65018ac176ee8608b0cb2a68e23c/share/lib/prim/Agda/Builtin/Bool.agda\"}, rangeFileName = Just (TopLevelModuleName {moduleNameRange = NoRange, moduleNameId = ModuleNameHash 4305008439024043551, moduleNameParts = \"Agda\" :| [\"Builtin\",\"Bool\"], moduleNameInferred = False})})) (fromList [Interval {getIntervalFile = (), iStart' = Pn {srcFile = (), posPos = 143, posLine = 4, posCol = 8}, iEnd' = Pn {srcFile = (), posPos = 148, posLine = 4, posCol = 13}}]), nameFixity = Fixity' {theFixity = Fixity {fixityRange = NoRange, fixityLevel = Unrelated, fixityAssoc = NonAssoc}, theNotation = [], theNameRange = NoRange}, nameIsRecordName = False},Name {nameId = NameId 2 (ModuleNameHash 4305008439024043551), nameConcrete = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"Builtin\" :| []}, nameCanonical = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"Builtin\" :| []}, nameBindingSite = Range (Just (RangeFile {rangeFilePath = AbsolutePath {textPath = \"/home/piter/.cabal/store/ghc-9.12.2-9b33/Agda-2.8.0-7cf9474c14d743910af4cb5c11fc608c8aef65018ac176ee8608b0cb2a68e23c/share/lib/prim/Agda/Builtin/Bool.agda\"}, rangeFileName = Just (TopLevelModuleName {moduleNameRange = NoRange, moduleNameId = ModuleNameHash 4305008439024043551, moduleNameParts = \"Agda\" :| [\"Builtin\",\"Bool\"], moduleNameInferred = False})})) (fromList [Interval {getIntervalFile = (), iStart' = Pn {srcFile = (), posPos = 148, posLine = 4, posCol = 13}, iEnd' = Pn {srcFile = (), posPos = 156, posLine = 4, posCol = 21}}]), nameFixity = Fixity' {theFixity = Fixity {fixityRange = NoRange, fixityLevel = Unrelated, fixityAssoc = NonAssoc}, theNotation = [], theNameRange = NoRange}, nameIsRecordName = False},Name {nameId = NameId 4 (ModuleNameHash 4305008439024043551), nameConcrete = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"Bool\" :| []}, nameCanonical = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"Bool\" :| []}, nameBindingSite = Range (Just (RangeFile {rangeFilePath = AbsolutePath {textPath = \"/home/piter/.cabal/store/ghc-9.12.2-9b33/Agda-2.8.0-7cf9474c14d743910af4cb5c11fc608c8aef65018ac176ee8608b0cb2a68e23c/share/lib/prim/Agda/Builtin/Bool.agda\"}, rangeFileName = Just (TopLevelModuleName {moduleNameRange = NoRange, moduleNameId = ModuleNameHash 4305008439024043551, moduleNameParts = \"Agda\" :| [\"Builtin\",\"Bool\"], moduleNameInferred = False})})) (fromList [Interval {getIntervalFile = (), iStart' = Pn {srcFile = (), posPos = 156, posLine = 4, posCol = 21}, iEnd' = Pn {srcFile = (), posPos = 160, posLine = 4, posCol = 25}}]), nameFixity = Fixity' {theFixity = Fixity {fixityRange = NoRange, fixityLevel = Unrelated, fixityAssoc = NonAssoc}, theNotation = [], theNameRange = NoRange}, nameIsRecordName = False}]}, qnameName = Name {nameId = NameId 6 (ModuleNameHash 4305008439024043551), nameConcrete = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"Bool\" :| []}, nameCanonical = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"Bool\" :| []}, nameBindingSite = Range (Just (RangeFile {rangeFilePath = AbsolutePath {textPath = \"/home/piter/.cabal/store/ghc-9.12.2-9b33/Agda-2.8.0-7cf9474c14d743910af4cb5c11fc608c8aef65018ac176ee8608b0cb2a68e23c/share/lib/prim/Agda/Builtin/Bool.agda\"}, rangeFileName = Just (TopLevelModuleName {moduleNameRange = NoRange, moduleNameId = ModuleNameHash 4305008439024043551, moduleNameParts = \"Agda\" :| [\"Builtin\",\"Bool\"], moduleNameInferred = False})})) (fromList [Interval {getIntervalFile = (), iStart' = Pn {srcFile = (), posPos = 173, posLine = 6, posCol = 6}, iEnd' = Pn {srcFile = (), posPos = 177, posLine = 6, posCol = 10}}]), nameFixity = Fixity' {theFixity = Fixity {fixityRange = NoRange, fixityLevel = Unrelated, fixityAssoc = NonAssoc}, theNotation = [], theNameRange = NoRange}, nameIsRecordName = False}}) []}))" */
+
+/* TODO In printScala2 got SeUnhandled "QName {qnameModule = MName {mnameToList = [Name {nameId = NameId 0 (ModuleNameHash 2665135241066394892), nameConcrete = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"examples\" :| []}, nameCanonical = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"examples\" :| []}, nameBindingSite = Range (Just (RangeFile {rangeFilePath = AbsolutePath {textPath = \"/home/piter/agda2scala/examples/higher_order_fun.agda\"}, rangeFileName = Just (TopLevelModuleName {moduleNameRange = NoRange, moduleNameId = ModuleNameHash 2665135241066394892, moduleNameParts = \"examples\" :| [\"higher_order_fun\"], moduleNameInferred = False})})) (fromList [Interval {getIntervalFile = (), iStart' = Pn {srcFile = (), posPos = 8, posLine = 1, posCol = 8}, iEnd' = Pn {srcFile = (), posPos = 17, posLine = 1, posCol = 17}}]), nameFixity = Fixity' {theFixity = Fixity {fixityRange = NoRange, fixityLevel = Unrelated, fixityAssoc = NonAssoc}, theNotation = [], theNameRange = NoRange}, nameIsRecordName = False},Name {nameId = NameId 2 (ModuleNameHash 2665135241066394892), nameConcrete = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"higher\" :| [Hole,Id \"order\",Hole,Id \"fun\"]}, nameCanonical = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"higher\" :| [Hole,Id \"order\",Hole,Id \"fun\"]}, nameBindingSite = Range (Just (RangeFile {rangeFilePath = AbsolutePath {textPath = \"/home/piter/agda2scala/examples/higher_order_fun.agda\"}, rangeFileName = Just (TopLevelModuleName {moduleNameRange = NoRange, moduleNameId = ModuleNameHash 2665135241066394892, moduleNameParts = \"examples\" :| [\"higher_order_fun\"], moduleNameInferred = False})})) (fromList [Interval {getIntervalFile = (), iStart' = Pn {srcFile = (), posPos = 17, posLine = 1, posCol = 17}, iEnd' = Pn {srcFile = (), posPos = 33, posLine = 1, posCol = 33}}]), nameFixity = Fixity' {theFixity = Fixity {fixityRange = NoRange, fixityLevel = Unrelated, fixityAssoc = NonAssoc}, theNotation = [], theNameRange = NoRange}, nameIsRecordName = False}]}, qnameName = Name {nameId = NameId 80 (ModuleNameHash 2665135241066394892), nameConcrete = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"with-78\" :| []}, nameCanonical = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"with-78\" :| []}, nameBindingSite = NoRange, nameFixity = Fixity' {theFixity = Fixity {fixityRange = NoRange, fixityLevel = Unrelated, fixityAssoc = NonAssoc}, theNotation = [], theNameRange = NoRange}, nameIsRecordName = False}}" "UnsupportedTerm (Pi (Dom {domInfo = ArgInfo {argInfoHiding = NotHidden, argInfoModality = Modality {modRelevance = Relevant ORelInferred, modQuantity = Quantity\969 Q\969Inferred, modCohesion = Continuous, modPolarity = PolarityModality {modPolarityAnn = MixedPolarity, modPolarityOrigin = MixedPolarity, modPolarityLock = StrictlyPositive}}, argInfoOrigin = UserWritten, argInfoFreeVariables = UnknownFVs, argInfoAnnotation = Annotation {annLock = IsNotLock}}, domName = Nothing, domIsFinite = False, domTactic = Nothing, unDom = El {_getSort = Univ UType (Max 0 []), unEl = Var 0 []}}) (NoAbs \"_\" El {_getSort = Univ UType (Max 0 []), unEl = Def (QName {qnameModule = MName {mnameToList = [Name {nameId = NameId 0 (ModuleNameHash 4305008439024043551), nameConcrete = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"Agda\" :| []}, nameCanonical = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"Agda\" :| []}, nameBindingSite = Range (Just (RangeFile {rangeFilePath = AbsolutePath {textPath = \"/home/piter/.cabal/store/ghc-9.12.2-9b33/Agda-2.8.0-7cf9474c14d743910af4cb5c11fc608c8aef65018ac176ee8608b0cb2a68e23c/share/lib/prim/Agda/Builtin/Bool.agda\"}, rangeFileName = Just (TopLevelModuleName {moduleNameRange = NoRange, moduleNameId = ModuleNameHash 4305008439024043551, moduleNameParts = \"Agda\" :| [\"Builtin\",\"Bool\"], moduleNameInferred = False})})) (fromList [Interval {getIntervalFile = (), iStart' = Pn {srcFile = (), posPos = 143, posLine = 4, posCol = 8}, iEnd' = Pn {srcFile = (), posPos = 148, posLine = 4, posCol = 13}}]), nameFixity = Fixity' {theFixity = Fixity {fixityRange = NoRange, fixityLevel = Unrelated, fixityAssoc = NonAssoc}, theNotation = [], theNameRange = NoRange}, nameIsRecordName = False},Name {nameId = NameId 2 (ModuleNameHash 4305008439024043551), nameConcrete = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"Builtin\" :| []}, nameCanonical = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"Builtin\" :| []}, nameBindingSite = Range (Just (RangeFile {rangeFilePath = AbsolutePath {textPath = \"/home/piter/.cabal/store/ghc-9.12.2-9b33/Agda-2.8.0-7cf9474c14d743910af4cb5c11fc608c8aef65018ac176ee8608b0cb2a68e23c/share/lib/prim/Agda/Builtin/Bool.agda\"}, rangeFileName = Just (TopLevelModuleName {moduleNameRange = NoRange, moduleNameId = ModuleNameHash 4305008439024043551, moduleNameParts = \"Agda\" :| [\"Builtin\",\"Bool\"], moduleNameInferred = False})})) (fromList [Interval {getIntervalFile = (), iStart' = Pn {srcFile = (), posPos = 148, posLine = 4, posCol = 13}, iEnd' = Pn {srcFile = (), posPos = 156, posLine = 4, posCol = 21}}]), nameFixity = Fixity' {theFixity = Fixity {fixityRange = NoRange, fixityLevel = Unrelated, fixityAssoc = NonAssoc}, theNotation = [], theNameRange = NoRange}, nameIsRecordName = False},Name {nameId = NameId 4 (ModuleNameHash 4305008439024043551), nameConcrete = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"Bool\" :| []}, nameCanonical = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"Bool\" :| []}, nameBindingSite = Range (Just (RangeFile {rangeFilePath = AbsolutePath {textPath = \"/home/piter/.cabal/store/ghc-9.12.2-9b33/Agda-2.8.0-7cf9474c14d743910af4cb5c11fc608c8aef65018ac176ee8608b0cb2a68e23c/share/lib/prim/Agda/Builtin/Bool.agda\"}, rangeFileName = Just (TopLevelModuleName {moduleNameRange = NoRange, moduleNameId = ModuleNameHash 4305008439024043551, moduleNameParts = \"Agda\" :| [\"Builtin\",\"Bool\"], moduleNameInferred = False})})) (fromList [Interval {getIntervalFile = (), iStart' = Pn {srcFile = (), posPos = 156, posLine = 4, posCol = 21}, iEnd' = Pn {srcFile = (), posPos = 160, posLine = 4, posCol = 25}}]), nameFixity = Fixity' {theFixity = Fixity {fixityRange = NoRange, fixityLevel = Unrelated, fixityAssoc = NonAssoc}, theNotation = [], theNameRange = NoRange}, nameIsRecordName = False}]}, qnameName = Name {nameId = NameId 6 (ModuleNameHash 4305008439024043551), nameConcrete = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"Bool\" :| []}, nameCanonical = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"Bool\" :| []}, nameBindingSite = Range (Just (RangeFile {rangeFilePath = AbsolutePath {textPath = \"/home/piter/.cabal/store/ghc-9.12.2-9b33/Agda-2.8.0-7cf9474c14d743910af4cb5c11fc608c8aef65018ac176ee8608b0cb2a68e23c/share/lib/prim/Agda/Builtin/Bool.agda\"}, rangeFileName = Just (TopLevelModuleName {moduleNameRange = NoRange, moduleNameId = ModuleNameHash 4305008439024043551, moduleNameParts = \"Agda\" :| [\"Builtin\",\"Bool\"], moduleNameInferred = False})})) (fromList [Interval {getIntervalFile = (), iStart' = Pn {srcFile = (), posPos = 173, posLine = 6, posCol = 6}, iEnd' = Pn {srcFile = (), posPos = 177, posLine = 6, posCol = 10}}]), nameFixity = Fixity' {theFixity = Fixity {fixityRange = NoRange, fixityLevel = Unrelated, fixityAssoc = NonAssoc}, theNotation = [], theNameRange = NoRange}, nameIsRecordName = False}}) []}))" */
+
+/* TODO In printScala2 got SeUnhandled "QName {qnameModule = MName {mnameToList = [Name {nameId = NameId 0 (ModuleNameHash 2665135241066394892), nameConcrete = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"examples\" :| []}, nameCanonical = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"examples\" :| []}, nameBindingSite = Range (Just (RangeFile {rangeFilePath = AbsolutePath {textPath = \"/home/piter/agda2scala/examples/higher_order_fun.agda\"}, rangeFileName = Just (TopLevelModuleName {moduleNameRange = NoRange, moduleNameId = ModuleNameHash 2665135241066394892, moduleNameParts = \"examples\" :| [\"higher_order_fun\"], moduleNameInferred = False})})) (fromList [Interval {getIntervalFile = (), iStart' = Pn {srcFile = (), posPos = 8, posLine = 1, posCol = 8}, iEnd' = Pn {srcFile = (), posPos = 17, posLine = 1, posCol = 17}}]), nameFixity = Fixity' {theFixity = Fixity {fixityRange = NoRange, fixityLevel = Unrelated, fixityAssoc = NonAssoc}, theNotation = [], theNameRange = NoRange}, nameIsRecordName = False},Name {nameId = NameId 2 (ModuleNameHash 2665135241066394892), nameConcrete = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"higher\" :| [Hole,Id \"order\",Hole,Id \"fun\"]}, nameCanonical = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"higher\" :| [Hole,Id \"order\",Hole,Id \"fun\"]}, nameBindingSite = Range (Just (RangeFile {rangeFilePath = AbsolutePath {textPath = \"/home/piter/agda2scala/examples/higher_order_fun.agda\"}, rangeFileName = Just (TopLevelModuleName {moduleNameRange = NoRange, moduleNameId = ModuleNameHash 2665135241066394892, moduleNameParts = \"examples\" :| [\"higher_order_fun\"], moduleNameInferred = False})})) (fromList [Interval {getIntervalFile = (), iStart' = Pn {srcFile = (), posPos = 17, posLine = 1, posCol = 17}, iEnd' = Pn {srcFile = (), posPos = 33, posLine = 1, posCol = 33}}]), nameFixity = Fixity' {theFixity = Fixity {fixityRange = NoRange, fixityLevel = Unrelated, fixityAssoc = NonAssoc}, theNotation = [], theNameRange = NoRange}, nameIsRecordName = False}]}, qnameName = Name {nameId = NameId 98 (ModuleNameHash 2665135241066394892), nameConcrete = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"maybe\" :| []}, nameCanonical = Name {nameRange = NoRange, nameInScope = InScope, nameNameParts = Id \"maybe\" :| []}, nameBindingSite = Range (Just (RangeFile {rangeFilePath = AbsolutePath {textPath = \"/home/piter/agda2scala/examples/higher_order_fun.agda\"}, rangeFileName = Just (TopLevelModuleName {moduleNameRange = NoRange, moduleNameId = ModuleNameHash 2665135241066394892, moduleNameParts = \"examples\" :| [\"higher_order_fun\"], moduleNameInferred = False})})) (fromList [Interval {getIntervalFile = (), iStart' = Pn {srcFile = (), posPos = 1056, posLine = 45, posCol = 1}, iEnd' = Pn {srcFile = (), posPos = 1061, posLine = 45, posCol = 6}}]), nameFixity = Fixity' {theFixity = Fixity {fixityRange = NoRange, fixityLevel = Unrelated, fixityAssoc = NonAssoc}, theNotation = [], theNameRange = NoRange}, nameIsRecordName = False}}" "UnsupportedTerm (Pi (Dom {domInfo = ArgInfo {argInfoHiding = NotHidden, argInfoModality = Modality {modRelevance = Relevant ORelInferred, modQuantity = Quantity\969 Q\969Inferred, modCohesion = Continuous, modPolarity = PolarityModality {modPolarityAnn = MixedPolarity, modPolarityOrigin = MixedPolarity, modPolarityLock = StrictlyPositive}}, argInfoOrigin = UserWritten, argInfoFreeVariables = UnknownFVs, argInfoAnnotation = Annotation {annLock = IsNotLock}}, domName = Nothing, domIsFinite = False, domTactic = Nothing, unDom = El {_getSort = Univ UType (Max 0 []), unEl = Var 2 []}}) (NoAbs \"_\" El {_getSort = Univ UType (Max 0 []), unEl = Var 1 []}))" */
+}
